MISUV-9576:   * minor code refactoring:: [renaming / add deprecation attributes where applicable];
