+++ app/controllers/FinancialDetailCreditsController.scala
@@ -20,7 +20,7 @@ import connectors.httpParsers.ChargeHttpParser.UnexpectedChargeResponse
 import controllers.predicates.AuthenticationPredicate
 import play.api.Logging
 import play.api.mvc._
-import services.FinancialDetailChargesService
+import services.FinancialDetailService
 import uk.gov.hmrc.play.bootstrap.backend.controller.BackendController
 import javax.inject.{Inject, Singleton}
 import scala.concurrent.ExecutionContext
@@ -28,7 +28,7 @@ import scala.concurrent.ExecutionContext
 @Singleton
 class FinancialDetailCreditsController @Inject()(authentication: AuthenticationPredicate,
                                                  cc: ControllerComponents,
-                                                 financialDetailChargesService : FinancialDetailChargesService)
+                                                 financialDetailChargesService : FinancialDetailService)
                                                 (implicit ec: ExecutionContext) extends BackendController(cc) with Logging {
 
   def getCredits(nino: String, fromDate: String, toDate: String): Action[AnyContent] = {
